djangorest/requirements/base.txt
alembic==0.9.4
autoenv==1.0.0
click==6.7
Django==1.11.7
django-environ==0.4.4
djangorestframework==3.7.1
Flask==0.12.2
Flask-API==0.7.1
Flask-Migrate==2.1.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.7
MarkupSafe==1.0
psycopg2==2.7.3
PyJWT==1.5.3
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.3
six==1.10.0
SQLAlchemy==1.1.13
Werkzeug==0.12.2
